+++ backend/src/Routify.Api/Models/LogModels/TextLogRowPayload.cs
@@ -11,8 +11,8 @@ public class TextLogRowPayload
     public int InputTokens { get; set; }
     public int OutputTokens { get; set; }
     
-    public double InputCost { get; set; }
-    public double OutputCost { get; set; }
+    public decimal InputCost { get; set; }
+    public decimal OutputCost { get; set; }
     
     public DateTime EndedAt { get; set; }
     public double Duration { get; set; }
